requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
PyPDF2==3.0.1
pydantic==2.5.0
requests==2.31.0
